move scraper to dedicated file
Decided against moving everything to their own dedicated package for now. Will probably be a good thing for the scraper eventually, but for now a dedicated file will do.

This should resolve #5, #6 for now.. When the need arises to create internal packages, a new issue will be created.
